Added beamline and location to mcd table. Refactored mcd_model to avoid circular imports
